init for 2.0
