+++ composer.json
@@ -10,19 +10,15 @@
     "description": "microservice framework base on swoole",
     "license": "Apache-2.0",
     "require": {
-        "swoft/framework": "^1.0"
     },
     "autoload": {
         "psr-4": {
             "Swoft\\View\\": "src/"
-        },
-        "files": [
-            "src/Helper/Functions.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
-            "Swoft\\View\\Test\\": "test"
+            "SwoftTest\\View\\": "test/case"
         }
     },
     "repositories": [
@@ -32,8 +28,6 @@
         }
     ],
     "require-dev": {
-        "swoft/swoole-ide-helper": "dev-master",
-        "phpunit/phpunit": "^5.7"
     },
     "scripts": {
         "test": "./vendor/bin/phpunit -c phpunit.xml"
