--- src/Base/AbstractViewInterface.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Swoft\View\Base;
-
-/**
- * the abstract class of veiw interface
- *
- * @uses      AbstractViewInterface
- * @version   2018年01月14日
- * @author    stelin <phpcrazy@126.com>
- * @copyright Copyright 2010-2016 swoft software
- * @license   PHP Version 7.x {@link http://www.php.net/license/3_0.txt}
- */
-abstract class AbstractViewInterface implements ViewInterface
-{
-}
