--- test/.env
@@ -1,50 +1,0 @@
-# test config
-TEST_NAME=test
-TEST_URI=127.0.0.1:6378
-TEST_MAX_IDEL=2
-TEST_MAX_ACTIVE=2
-TEST_MAX_WAIT=2
-TEST_TIMEOUT=2
-TEST_BALANCER=r1
-TEST_USE_PROVIDER=true
-TEST_PROVIDER=c1
-
-# the pool of master nodes pool
-DB_NAME=master2
-DB_URI=127.0.0.1:3302,127.0.0.1:3302
-DB_MAX_IDEL=2
-DB_MAX_ACTIVE=2
-DB_MAX_WAIT=2
-DB_TIMEOUT=2
-DB_USE_PROVIDER=true
-DB_BALANCER=random2
-DB_PROVIDER=consul2
-
-# the pool of slave nodes pool
-DB_SLAVE_NAME=slave2
-DB_SLAVE_URI=127.0.0.1:3306/test?user=root&password=&charset=utf8,127.0.0.1:3306/test?user=root&password=&charset=utf8
-DB_SLAVE_MAX_IDEL=2
-DB_SLAVE_MAX_ACTIVE=2
-DB_SLAVE_MAX_WAIT=2
-DB_SLAVE_TIMEOUT=2
-DB_SLAVE_USE_PROVIDER=false
-DB_SLAVE_BALANCER=random
-DB_SLAVE_PROVIDER=consul2
-
-# the pool of redis
-REDIS_NAME=redis2
-REDIS_URI=127.0.0.1:2222,127.0.0.1:2222
-REDIS_MAX_IDEL=2
-REDIS_MAX_ACTIVE=2
-REDIS_MAX_WAIT=2
-REDIS_TIMEOUT=2
-REDIS_USE_PROVIDER=true
-REDIS_BALANCER=random2
-REDIS_PROVIDER=consul2
-
-
-# consul provider
-PROVIDER_CONSUL_ADDRESS=http://127.0.0.1:82
-PROVIDER_CONSUL_TAGS=1,2
-PROVIDER_CONSUL_TIMEOUT=2
-PROVIDER_CONSUL_INTERVAL=2
