--- test/config/beans/base.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'eventManager'    => [
-        'class'     => \Swoft\Event\EventManager::class,
-    ],
-];
