--- test/config/properties/app.php
@@ -1,19 +1,0 @@
-<?php
-return [
-    "version"           => '1.0',
-    'autoInitBean'      => true,
-    'beanScan'          => [
-        'Swoft\\View\\Test\\Testing' => BASE_PATH."/Testing"
-    ],
-    'I18n'              => [
-        'sourceLanguage' => '@root/resources/messages/',
-    ],
-    'env'               => 'Base',
-    'user.stelin.steln' => 'fafafa',
-    'Service'           => [
-        'user' => [
-            'timeout' => 3000
-        ]
-    ],
-    'cache' => require dirname(__FILE__) . DS . "cache.php",
-];
